# Remove device properties from miners