--- a/Include.psm1
+++ b/Include.psm1
@@ -741,8 +741,6 @@ class Miner {
     [Array]$Algorithm = @()
     $Type
     $Index
-    $Device
-    $Device_Auto
     $Profit
     $Profit_Comparison
     $Profit_MarginOfError