This is the single largest performance improvement I've found yet.  It saves 30-50s on every loop on my fastest system, and about 1 and a half minutes on my slowest.

The device_auto code was added quite awhile ago, in https://github.com/MultiPoolMiner/MultiPoolMiner/commit/8cba63210b0691841b973283bd420329a524f188, and as best as I can tell, was only really used by Excavator. Excavator has since been rewritten a few times - the devices part was removed in https://github.com/MultiPoolMiner/MultiPoolMiner/commit/9ca604a6bf8ee94d17ff4ebe17a0a5469c0ff882

None of the miners use this feature anymore, and any per-device mining that might be added will be done with the existing Index property.  So this code isn't actually being used anymore, and removing it doesn't seem to affect the profit calculations at all.

## Benchmarks running a single loop:
These tests were run on the current master, with all miners and algorithms enabled, and all benchmarking finished.

|Test|System|Before|After|Savings|
|----|------|------|-----|-------|
|1|Fast (i7)|2:04|1:28|36s|
|2|Fast (i7)|2:13|1:29|44s|
|1|Slow (AMD A10)|4:44|3:18|86s|
|2|Slow (AMD A10)|5:25|3:42|103s|