@aaronsace 2018-05-14
Everything should be moving to device i.e. 'CPU#01', 'GPU#01', 'GPU#02'. Type and Index should be legacy. I've mentioned a 'Get-Device' function to help the miner files before; please ensure you are directing towards this.