@grantemsley 2018-05-14
@aaronsace I don't recall the Get-Device function you're referring to.  Do you remember where you mentioned it?

The current Type and Index will definitely get changed as we move towards being able to run miners on a per-device basis instead of per-type.