@aaronsace 2018-05-15
It's definitely come up in conversation somewhere. I had an implementation of Get-Device but I can't find it; it's a simple function but I couldn't confirm if the order matched task manger. 

This is how it works: 
Device should match task manager if possible; if not possible, it should be the OpenCL order. 
Get-Device should return the Devices with the 'name' (GPU#01, GPU#02 etc.) plus various fields (i.e. platform, index etc.) that each miner file would use to sort their fields to allow them to use it within the command line and a 'device' parameter (instead of 'type' and 'index' parameter). 
A new 'device' filter should be added to the core code command line and the 'type' filter should be added as an alias to the 'device' parameter. The 'device' filter should be flexible so that it excepts not only things like 'GPU#01' and 'CPU#01' but also things like 'NVidia', 'GPU', 'CPU', 'AMD' etc.